+++ src/components/pages/admin/callouts/steps/ContentStep.vue
@@ -33,10 +33,10 @@ import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRoute } from 'vue-router';
 import { ContentStepProps } from '../callouts.interface';
-import RichTextEditor from '../../../rte/RichTextEditor.vue';
-import AppFormSection from '../../../forms/AppFormSection.vue';
+import RichTextEditor from '../../../../rte/RichTextEditor.vue';
+import AppFormSection from '../../../../forms/AppFormSection.vue';
 
-import AppNotification from '../../../AppNotification.vue';
+import AppNotification from '../../../../AppNotification.vue';
 import ContentForm from './ContentForm.vue';
 
 const emit = defineEmits(['update:error', 'update:validated']);
