+++ src/components/pages/admin/callouts/steps/DatesAndDuration.vue
@@ -50,10 +50,10 @@ import { ItemStatus } from '@beabee/beabee-common';
 import useVuelidate from '@vuelidate/core';
 import { computed, ref, toRef, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import AppInput from '../../../forms/AppInput.vue';
-import AppLabel from '../../../forms/AppLabel.vue';
-import AppRadioGroup from '../../../forms/AppRadioGroup.vue';
-import AppFormSection from '../../../forms/AppFormSection.vue';
+import AppInput from '../../../../forms/AppInput.vue';
+import AppLabel from '../../../../forms/AppLabel.vue';
+import AppRadioGroup from '../../../../forms/AppRadioGroup.vue';
+import AppFormSection from '../../../../forms/AppFormSection.vue';
 import { DateAndDurationStepProps } from '../callouts.interface';
 import { sameAs } from '@vuelidate/validators';
