+++ src/components/pages/admin/callouts/steps/EndMessage.vue
@@ -47,11 +47,11 @@
 import useVuelidate from '@vuelidate/core';
 import { watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import AppInput from '../../../forms/AppInput.vue';
-import AppRadioGroup from '../../../forms/AppRadioGroup.vue';
+import AppInput from '../../../../forms/AppInput.vue';
+import AppRadioGroup from '../../../../forms/AppRadioGroup.vue';
 import { EndMessageStepProps } from '../callouts.interface';
-import RichTextEditor from '../../../rte/RichTextEditor.vue';
-import AppFormSection from '../../../forms/AppFormSection.vue';
+import RichTextEditor from '../../../../rte/RichTextEditor.vue';
+import AppFormSection from '../../../../forms/AppFormSection.vue';
 
 const emit = defineEmits(['update:error', 'update:validated']);
 defineProps<{ data: EndMessageStepProps }>();
