+++ src/pages/admin/callouts/new.vue
@@ -37,7 +37,7 @@ meta:
         </AppAsyncButton>
       </div>
     </PageTitle>
-    <CalloutForm :steps-props="steps" :status="status" />
+    <CalloutStepper :steps-props="steps" :status="status" />
   </div>
 </template>
 
@@ -51,8 +51,8 @@ import {
   fetchCallout,
   updateCallout,
 } from '../../../utils/api/callout';
-import { CalloutStepsProps } from '../../../components/pages/callouts/callouts.interface';
-import CalloutForm from '../../../components/pages/callouts/CalloutForm.vue';
+import { CalloutStepsProps } from '../../../components/pages/admin/callouts/callouts.interface';
+import CalloutStepper from '../../../components/pages/admin/callouts/CalloutStepper.vue';
 import {
   convertCalloutToSteps,
   convertStepsToCallout,
